tests/functional/requirements.txt
pydantic-settings==2.2.1
kafka-python==2.0.2
python-dotenv==1.0.1
aiokafka==0.8.1
aiohttp==3.9.3
pytest==7.4.4
pytest-asyncio==0.12.0
pydantic==2.3.0
starlette==0.27.0